+++ java/adapter/avro/src/test/java/org/apache/arrow/AvroToArrowTest.java
@@ -26,18 +26,14 @@ import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Set;
 
 import org.apache.arrow.vector.FieldVector;
 import org.apache.arrow.vector.VarCharVector;
 import org.apache.arrow.vector.VectorSchemaRoot;
 import org.apache.arrow.vector.complex.ListVector;
 import org.apache.arrow.vector.complex.MapVector;
-import org.apache.arrow.vector.complex.StructVector;
-import org.apache.arrow.vector.types.Types;
 import org.apache.avro.Schema;
 import org.apache.avro.generic.GenericData;
 import org.apache.avro.generic.GenericDatumWriter;
@@ -127,259 +123,6 @@ public class AvroToArrowTest extends AvroTestBase {
   }
 
   @Test
-  public void testSkipUnionWithOneField() throws Exception {
-    Set<String> skipFieldNames = new HashSet<>();
-    skipFieldNames.add("f0");
-    config = new AvroToArrowConfigBuilder(config.getAllocator()).setSkipFieldNames(skipFieldNames).build();
-    Schema schema = getSchema("skip/test_skip_union_before.avsc");
-    Schema expectedSchema = getSchema("skip/test_skip_union_one_field_expected.avsc");
-
-    ArrayList<GenericRecord> data = new ArrayList<>();
-    ArrayList<GenericRecord> expectedData = new ArrayList<>();
-
-    for (int i = 0; i < 5; i++) {
-      GenericRecord record = new GenericData.Record(schema);
-      record.put(0, "test" + i);
-      record.put(1, i % 2 == 0 ? "test" + i : null);
-      record.put(2, i % 2 == 0 ? "test" + i : i);
-      record.put(3, i);
-      data.add(record);
-
-      GenericRecord expectedRecord = new GenericData.Record(expectedSchema);
-      expectedRecord.put(0, record.get(1));
-      expectedRecord.put(1, record.get(2));
-      expectedRecord.put(2, record.get(3));
-      expectedData.add(expectedRecord);
-    }
-    VectorSchemaRoot root = writeAndRead(schema, data);
-    checkRecordResult(expectedSchema, expectedData, root);
-  }
-
-  @Test
-  public void testSkipUnionWithNullableOneField() throws Exception {
-    Set<String> skipFieldNames = new HashSet<>();
-    skipFieldNames.add("f1");
-    config = new AvroToArrowConfigBuilder(config.getAllocator()).setSkipFieldNames(skipFieldNames).build();
-    Schema schema = getSchema("skip/test_skip_union_before.avsc");
-    Schema expectedSchema = getSchema("skip/test_skip_union_nullable_field_expected.avsc");
-
-    ArrayList<GenericRecord> data = new ArrayList<>();
-    ArrayList<GenericRecord> expectedData = new ArrayList<>();
-
-    for (int i = 0; i < 5; i++) {
-      GenericRecord record = new GenericData.Record(schema);
-      record.put(0, "test" + i);
-      record.put(1, i % 2 == 0 ? "test" + i : null);
-      record.put(2, i % 2 == 0 ? "test" + i : i);
-      record.put(3, i);
-      data.add(record);
-
-      GenericRecord expectedRecord = new GenericData.Record(expectedSchema);
-      expectedRecord.put(0, record.get(0));
-      expectedRecord.put(1, record.get(2));
-      expectedRecord.put(2, record.get(3));
-      expectedData.add(expectedRecord);
-    }
-    VectorSchemaRoot root = writeAndRead(schema, data);
-    checkRecordResult(expectedSchema, expectedData, root);
-  }
-
-  @Test
-  public void testSkipUnionWithMultiFields() throws Exception {
-    Set<String> skipFieldNames = new HashSet<>();
-    skipFieldNames.add("f2");
-    config = new AvroToArrowConfigBuilder(config.getAllocator()).setSkipFieldNames(skipFieldNames).build();
-    Schema schema = getSchema("skip/test_skip_union_before.avsc");
-    Schema expectedSchema = getSchema("skip/test_skip_union_multi_fields_expected.avsc");
-
-    ArrayList<GenericRecord> data = new ArrayList<>();
-    ArrayList<GenericRecord> expectedData = new ArrayList<>();
-
-    for (int i = 0; i < 5; i++) {
-      GenericRecord record = new GenericData.Record(schema);
-      record.put(0, "test" + i);
-      record.put(1, i % 2 == 0 ? "test" + i : null);
-      record.put(2, i % 2 == 0 ? "test" + i : i);
-      record.put(3, i);
-      data.add(record);
-
-      GenericRecord expectedRecord = new GenericData.Record(expectedSchema);
-      expectedRecord.put(0, record.get(0));
-      expectedRecord.put(1, record.get(1));
-      expectedRecord.put(2, record.get(3));
-      expectedData.add(expectedRecord);
-    }
-    VectorSchemaRoot root = writeAndRead(schema, data);
-    checkRecordResult(expectedSchema, expectedData, root);
-  }
-
-  @Test
-  public void testSkipArrayField() throws Exception {
-    Set<String> skipFieldNames = new HashSet<>();
-    skipFieldNames.add("f1");
-    config = new AvroToArrowConfigBuilder(config.getAllocator()).setSkipFieldNames(skipFieldNames).build();
-    Schema schema = getSchema("skip/test_skip_array_before.avsc");
-    Schema expectedSchema = getSchema("skip/test_skip_array_expected.avsc");
-
-    ArrayList<GenericRecord> data = new ArrayList<>();
-    ArrayList<GenericRecord> expectedData = new ArrayList<>();
-
-    for (int i = 0; i < 5; i++) {
-      GenericRecord record = new GenericData.Record(schema);
-      record.put(0, "test" + i);
-      record.put(1, Arrays.asList("test" + i, "test" + i));
-      record.put(2, i % 2 == 0);
-      data.add(record);
-
-      GenericRecord expectedRecord = new GenericData.Record(expectedSchema);
-      expectedRecord.put(0, record.get(0));
-      expectedRecord.put(1, record.get(2));
-      expectedData.add(expectedRecord);
-    }
-
-    VectorSchemaRoot root = writeAndRead(schema, data);
-    checkRecordResult(expectedSchema, expectedData, root);
-  }
-
-  @Test
-  public void testSkipMultiFields() throws Exception {
-    Set<String> skipFieldNames = new HashSet<>();
-    skipFieldNames.add("f1");
-    skipFieldNames.add("f2");
-    config = new AvroToArrowConfigBuilder(config.getAllocator()).setSkipFieldNames(skipFieldNames).build();
-    Schema schema = getSchema("test_record.avsc");
-    Schema expectedSchema = getSchema("skip/test_skip_multi_fields_expected.avsc");
-
-    ArrayList<GenericRecord> data = new ArrayList<>();
-    ArrayList<GenericRecord> expectedData = new ArrayList<>();
-
-    for (int i = 0; i < 5; i++) {
-      GenericRecord record = new GenericData.Record(schema);
-      record.put(0, "test" + i);
-      record.put(1, i);
-      record.put(2, i % 2 == 0);
-      data.add(record);
-
-      GenericRecord expectedRecord = new GenericData.Record(expectedSchema);
-      expectedRecord.put(0, record.get(0));
-      expectedData.add(expectedRecord);
-    }
-
-    VectorSchemaRoot root = writeAndRead(schema, data);
-    checkRecordResult(expectedSchema, expectedData, root);
-  }
-
-  @Test
-  public void testSkipField() throws Exception {
-    Set<String> skipFieldNames = new HashSet<>();
-    skipFieldNames.add("f1");
-    config = new AvroToArrowConfigBuilder(config.getAllocator()).setSkipFieldNames(skipFieldNames).build();
-    Schema schema = getSchema("test_record.avsc");
-    Schema expectedSchema = getSchema("skip/test_skip_single_field_expected.avsc");
-
-    ArrayList<GenericRecord> data = new ArrayList<>();
-    ArrayList<GenericRecord> expectedData = new ArrayList<>();
-
-    for (int i = 0; i < 5; i++) {
-      GenericRecord record = new GenericData.Record(schema);
-      record.put(0, "test" + i);
-      record.put(1, i);
-      record.put(2, i % 2 == 0);
-      data.add(record);
-
-      GenericRecord expectedRecord = new GenericData.Record(expectedSchema);
-      expectedRecord.put(0, record.get(0));
-      expectedRecord.put(1, record.get(2));
-      expectedData.add(expectedRecord);
-    }
-
-    VectorSchemaRoot root = writeAndRead(schema, data);
-    checkRecordResult(expectedSchema, expectedData, root);
-  }
-
-  @Test
-  public void testSkipNestedFields() throws Exception {
-    Set<String> skipFieldNames = new HashSet<>();
-    skipFieldNames.add("f0.f0");
-    config = new AvroToArrowConfigBuilder(config.getAllocator()).setSkipFieldNames(skipFieldNames).build();
-    Schema schema = getSchema("test_nested_record.avsc");
-    Schema nestedSchema = schema.getFields().get(0).schema();
-    ArrayList<GenericRecord> data = new ArrayList<>();
-
-    Schema expectedSchema = getSchema("skip/test_skip_second_level_expected.avsc");
-    Schema expectedNestedSchema = expectedSchema.getFields().get(0).schema();
-    ArrayList<GenericRecord> expectedData = new ArrayList<>();
-
-    for (int i = 0; i < 5; i++) {
-      GenericRecord record = new GenericData.Record(schema);
-      GenericRecord nestedRecord = new GenericData.Record(nestedSchema);
-      nestedRecord.put(0, "test" + i);
-      nestedRecord.put(1, i);
-      record.put(0, nestedRecord);
-      data.add(record);
-
-      GenericRecord expectedRecord = new GenericData.Record(expectedSchema);
-      GenericRecord expectedNestedRecord = new GenericData.Record(expectedNestedSchema);
-      expectedNestedRecord.put(0, nestedRecord.get(1));
-      expectedRecord.put(0, expectedNestedRecord);
-      expectedData.add(expectedRecord);
-    }
-
-    VectorSchemaRoot root = writeAndRead(schema, data);
-    checkNestedRecordResult(expectedSchema, expectedData, root);
-  }
-
-  @Test
-  public void testSkipThirdLevelField() throws Exception {
-    Schema firstLevelSchema = getSchema("skip/test_skip_third_level_expected.avsc");
-    Schema secondLevelSchema = firstLevelSchema.getFields().get(0).schema();
-    Schema thirdLevelSchema = secondLevelSchema.getFields().get(0).schema();
-
-    ArrayList<GenericRecord> data = new ArrayList<>();
-    for (int i = 0; i < 5; i++) {
-      GenericRecord firstLevelRecord = new GenericData.Record(firstLevelSchema);
-      GenericRecord secondLevelRecord = new GenericData.Record(secondLevelSchema);
-      GenericRecord thirdLevelRecord = new GenericData.Record(thirdLevelSchema);
-
-      thirdLevelRecord.put(0, i);
-      thirdLevelRecord.put(1, "test" + i);
-      thirdLevelRecord.put(2, i % 2 == 0);
-
-      secondLevelRecord.put(0, thirdLevelRecord);
-      firstLevelRecord.put(0, secondLevelRecord);
-      data.add(firstLevelRecord);
-    }
-
-    // do not skip any fields first
-    VectorSchemaRoot root1 = writeAndRead(firstLevelSchema, data);
-
-    assertEquals(1, root1.getFieldVectors().size());
-    assertEquals(Types.MinorType.STRUCT, root1.getFieldVectors().get(0).getMinorType());
-    StructVector secondLevelVector = (StructVector) root1.getFieldVectors().get(0);
-    assertEquals(1, secondLevelVector.getChildrenFromFields().size());
-    assertEquals(Types.MinorType.STRUCT, secondLevelVector.getChildrenFromFields().get(0).getMinorType());
-    StructVector thirdLevelVector = (StructVector) secondLevelVector.getChildrenFromFields().get(0);
-    assertEquals(3, thirdLevelVector.getChildrenFromFields().size());
-
-    // skip third level field and validate
-    Set<String> skipFieldNames = new HashSet<>();
-    skipFieldNames.add("f0.f0.f0");
-    config = new AvroToArrowConfigBuilder(config.getAllocator()).setSkipFieldNames(skipFieldNames).build();
-    VectorSchemaRoot root2 = writeAndRead(firstLevelSchema, data);
-
-    assertEquals(1, root2.getFieldVectors().size());
-    assertEquals(Types.MinorType.STRUCT, root2.getFieldVectors().get(0).getMinorType());
-    StructVector secondStruct = (StructVector) root2.getFieldVectors().get(0);
-    assertEquals(1, secondStruct.getChildrenFromFields().size());
-    assertEquals(Types.MinorType.STRUCT, secondStruct.getChildrenFromFields().get(0).getMinorType());
-    StructVector thirdStruct = (StructVector) secondStruct.getChildrenFromFields().get(0);
-    assertEquals(2, thirdStruct.getChildrenFromFields().size());
-
-    assertEquals(Types.MinorType.INT, thirdStruct.getChildrenFromFields().get(0).getMinorType());
-    assertEquals(Types.MinorType.BIT, thirdStruct.getChildrenFromFields().get(1).getMinorType());
-  }
-
-  @Test
   public void testEnumType() throws Exception {
     Schema schema = getSchema("test_primitive_enum.avsc");
     List<GenericData.EnumSymbol> data = Arrays.asList(
@@ -389,7 +132,7 @@ public class AvroToArrowTest extends AvroTestBase {
         new GenericData.EnumSymbol(schema, "CLUBS"),
         new GenericData.EnumSymbol(schema, "SPADES"));
 
-    List<Byte> expectedIndices = Arrays.asList((byte)0, (byte)1, (byte)2, (byte)3, (byte)0);
+    List<Integer> expectedIndices = Arrays.asList(0, 1, 2, 3, 0);
 
     VectorSchemaRoot root = writeAndRead(schema, data);
     FieldVector vector = root.getFieldVectors().get(0);
