+++ java/adapter/avro/src/test/resources/schema/skip/test_skip_record_before.avsc
@@ -1,0 +1,38 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+{
+ "namespace": "org.apache.arrow.avro",
+ "type": "record",
+ "name": "test",
+ "fields": [
+  {
+    "name" : "f0",
+    "type" : {
+        "type" : "record",
+        "name" : "nestedInRecord",
+        "fields": [
+             {"name": "f00", "type": "string"},
+             {"name": "f01", "type": "int"}
+        ]
+    }
+  },
+  {
+    "name" : "f1", "type" : "int"
+  }
+  ]
+}
